all the integer conversions (%u, %i) are now handeled with the length modifier (h, hh, l, ll, j, z and t) and the flags - and 0

+++ libft/srcs/ft_freturn.c
@@ -6,13 +6,13 @@
 /*   By: abaurens <abaurens@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/04 01:52:23 by abaurens          #+#    #+#             */
-/*   Updated: 2018/12/04 01:55:15 by abaurens         ###   ########.fr       */
+/*   Updated: 2018/12/13 18:48:32 by abaurens         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 
-int		ft_freturn(void *ptr, int ret)
+long long int	ft_freturn(void *ptr, long long int ret)
 {
 	free(ptr);
 	return (ret);
